consumer/pyproject.toml
[tool.poetry]
name = "consumer"
version = "0.1.0"
description = ""
authors = ["lijian418 <126387358+lijian418@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
motor = "^3.1.1"
python-dotenv = "^1.0.0"
unicode = "^2.9"
six = "^1.16.0"
aio-pika = "^9.0.4"
pika = "^1.3.1"
pydantic = "^1.10.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
